// src/app/core/models/masters/human-capital/designation-zone.ts
export interface DesignationZone {
    id: number;
    title: string;
    status: string;
    status_id: number;
    status_label: string;
}
export interface DesignationZonePaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: DesignationZone[];
}